microservices/user-service/requirements.txt
flask==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
bcrypt==4.0.1
pyjwt==2.6.0
